Increase speed and robustness of tmux/zsh plugins
Reduced the total number of tests and added and explicit test to make
sure the the hostname is not duplicated in the title.

Also changed SSH detection to use only SSH_CONNECTION and the parent
PID, which allows for the SSH status to be refreshed depending on
whether the user is currently accessing SSH or not.